Add support for merging unique psbts
This operation is needed for NS1R payjoins and in the future for the
more generalized NSMR case. Rust bitcoin does support combining PSBTs
that include the exact same unsigned tx; however, it does not have
support for merging PSBTs that are unique.

We add the additional merge operation as a new trait extension under the
merge submodule of a newly created psbt module.