Persist receiver when initialized
This commit introduces a new `Persister` parameter to
`v2::Receiver::new`, making it explicit that persistence is required
when creating a receiver. Applications can determine how and in what
format to persist the receiver state by implementing the `Persistable`
trait.

The `Persistable` key for the receiver is the session `ShortId`, which,
as of this commit, matches the key used in the Payjoin CLI.

Issue: https://github.com/payjoin/rust-payjoin/issues/336