Persist sender when initialized
This commit adds a new `Persister` parameter to `v2::Sender::new`,
making it explicit that persistence is required when creating a
sender. Applications can define how and in what format to persist the
sender state by implementing the `Persistable` trait. The
`Persistable` key for the sender is the directory ID associated with a
particular session, aligning with the key used by Payjoin CLI as of this
commit.

For cases where persistence is not needed, such as testing, a no-op
persister is available.

Issue: https://github.com/payjoin/rust-payjoin/issues/336