Introduce `ReceiverSessionEvent`
This commit introduces session events created by processing the receiver
state machine. These generally wrap v1 receiver types as not to
duplicate session context which is learned by handling the first typestate.

In order to support serializing and deserializing the session events we
need to impl `serde::de/serialize` on the underlying v1 types. And as a
side effect this means we need to enable `bitcoin/serde` under the v1
flag since `bitcoin::FeeRate` is a field under `OptionalParameters`.